requirements.txt
flask
flask-login
pandas
openpyxl
requests
gunicorn
gspread
oauth2client
python-dotenv
SQLAlchemy
psycopg2-binary
reportlab